refactor(feat): 基于官方microsoft的script启动器示例项目重构项目框架底座
