+++ .env
@@ -1,0 +1,3 @@
+PROJECT_NAME="usf"
+MINECRAFT_PRODUCT="BedrockGDK"
+CUSTOM_DEPLOYMENT_PATH=""
