+++ .gitignore
@@ -1,1 +1,2 @@
 *.bak
+node_modules/
