--- BehaviourPack/entities/small_fireball.json
@@ -1,54 +1,0 @@
-{
-  "format_version": "1.12.0",
- "minecraft:entity": {
-  "description": {
-    "identifier": "minecraft:small_fireball",
-    "is_spawnable": false,
-    "is_summonable": true,
-    "is_experimental": false
-  },
-  "components": {
-   "minecraft:collision_box": {
-    "width": 0.31,
-    "height": 0.31
-   },
-   "minecraft:projectile": {
-
-    "on_hit": {
-      "impact_damage": {
-        "damage": 5,
-        "knockback": true,
-        "catch_fire": true,
-        "semi_random_diff_damage": false
-      },
-      "catch_fire": {
-        "fire_affected_by_griefing": true
-      },
-      "remove_on_hit": { }
-    },
-    "power": 1.3,
-    "gravity": 0.00,
-    "inertia": 1,
-    "liquid_inertia": 1,
-    "anchor": 2,
-    "offset": [ 0, 0.5, 0 ],
-    "semi_random_diff_damage": true,
-    "uncertainty_base": 10.0,
-    "reflect_on_hurt": true
-   },
-   "minecraft:physics": {
-   },
-   "minecraft:pushable": {
-    "is_pushable": true,
-    "is_pushable_by_piston": true
-   },
-   "minecraft:conditional_bandwidth_optimization": {
-    "default_values": {
-      "max_optimized_distance": 80.0,
-      "max_dropped_ticks": 7,
-      "use_motion_prediction_hints": true
-    }
-  }
-  }
- }
-}
