--- LogServer/LogServer.json
@@ -1,4 +1,0 @@
-{
-    "address" : "127.0.0.1",
-    "port" : 1024 
-}
