--- LogServer/permissions.json
@@ -1,10 +1,0 @@
-{
-  "allowed_modules": [
-    "@minecraft/server-gametest",
-    "@minecraft/server",
-    "@minecraft/server-ui",
-    "@minecraft/server-admin",
-    "@minecraft/server-editor",
-	"@minecraft/server-net"
-  ]
-}
