--- LogServer/start.cmd
@@ -1,2 +1,0 @@
-chcp 65001
-PowerShell -Command "node Main.js"
