--- LogServer/start.sh
@@ -1,1 +1,0 @@
-node ./Main.js
