+++ behavior_packs/usf/manifest.json
@@ -14,7 +14,7 @@
             "language":"javascript",
             "uuid": "317b0654-acf7-903f-8639-9dd226ee395f",
             "version": [0, 7, 9],
-            "entry": "scripts/Main.js"
+            "entry": "scripts/main.js"
         }
     ],
     "dependencies": [
