+++ log-server/log-server.json
@@ -1,0 +1,4 @@
+{
+    "address" : "127.0.0.1",
+    "port" : 1024 
+}
