+++ log-server/start.cmd
@@ -1,0 +1,2 @@
+chcp 65001
+PowerShell -Command "node main.js"
