+++ log-server/start.sh
@@ -1,0 +1,1 @@
+node ./main.js
