+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "USF",
+  "version": "0.8.0",
+  "productName": "A minecraft bds Framework",
+  "description": "A minecraft bds Framework",
+  "private": true,
+  "devDependencies": {
+    "@minecraft/core-build-tasks": "^5.5.0",
+    "eslint-plugin-minecraft-linting": "^2.0.10",
+    "source-map": "^0.7.4",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.5.4"
+  },
+  "scripts": {
+    "lint": "just-scripts lint",
+    "build": "just-scripts build",
+    "build:production": "just-scripts build --production",
+    "clean": "just-scripts clean",
+    "local-deploy": "just-scripts local-deploy",
+    "mcaddon": "just-scripts mcaddon",
+    "mcaddon:production": "just-scripts mcaddon --production",
+    "enablemcloopback": "CheckNetIsolation.exe LoopbackExempt -a -p=S-1-15-2-1958404141-86561845-1752920682-3514627264-368642714-62675701-733520436",
+    "enablemcpreviewloopback": "CheckNetIsolation.exe LoopbackExempt -a -p=S-1-15-2-424268864-5579737-879501358-346833251-474568803-887069379-4040235476"
+  },
+  "dependencies": {
+    "@minecraft/math": "2.4.0",
+    "@minecraft/server": "2.9.0-beta.1.26.32-stable",
+    "@minecraft/server-ui": "2.2.0-beta.1.26.32-stable",
+    "@minecraft/vanilla-data": "1.26.32"
+  }
+}
