+++ scripts/functions/get_owner.mcfunction
