+++ scripts/functions/reset.mcfunction
