+++ scripts/main.ts
@@ -1,0 +1,1 @@
+// 入口脚本
