+++ scripts/scripts/Main.js
