+++ scripts/scripts/text.js
