+++ tsconfig.json
@@ -1,0 +1,31 @@
+{ 
+    "compilerOptions":{ 
+       "target":"es6",
+       "moduleResolution":"Node",
+       "module": "ES2020",
+       "declaration":false,
+       "noLib":false,
+       "emitDecoratorMetadata":true,
+       "experimentalDecorators":true,
+       "sourceMap":true,
+       "pretty":true,
+       "forceConsistentCasingInFileNames": true,
+       "strict": true,
+       "allowUnreachableCode":true,
+       "allowUnusedLabels":true,
+       "noImplicitAny":true,
+       "noImplicitReturns":false,
+       "noImplicitUseStrict":false,
+       "outDir":"lib",
+       "rootDir": ".",
+       "baseUrl":"behavior_packs/",
+       "listFiles":false,
+       "noEmitHelpers":true,
+       "skipLibCheck": true
+    },
+    "include":[ 
+       "scripts/**/*"
+    ],
+    "exclude": ["lib", "dist", "node_modules"],
+    "compileOnSave":false
+ }
